apps/api/dependencies/security.py
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBearer
from datetime import datetime, timedelta
from jose import JWTError, jwt
from .config import config
from .database import get_db
from sqlalchemy.ext.asyncio import AsyncSession
from models import User
from sqlmodel import select

security = HTTPBearer()

def create_access_token(data: dict, expires_delta: timedelta = None):
    to_encode = data.copy()
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=config.ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, config.SECRET_KEY, algorithm=config.ALGORITHM)

def verify_access_token(token: str):
    try:
        payload = jwt.decode(token, config.SECRET_KEY, algorithms=[config.ALGORITHM])
        email: str = payload.get("sub")
        if email is None:
            raise JWTError
        return email
    except JWTError:
        raise JWTError

async def get_current_user(db: AsyncSession = Depends(get_db), token: str = Depends(security)):
    credentials_exception = HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Could not validate credentials")
    try:
        email = verify_access_token(token.credentials)
        user = (await db.exec(select(User).where(User.email == email))).first()
        if user is None:
            raise credentials_exception
        return user
    except:
        raise credentials_exception